+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - (Feature) Add action progress
 - (Feature) Ensure consistency during replication cancellation
 - (Feature) Add annotation to change architecture of a member
+- (Bugfix) Prevent Member Maintenance Error log
 
 ## [1.2.18](https://github.com/arangodb/kube-arangodb/tree/1.2.18) (2022-09-28)
 - (Feature) Define Actions PlaceHolder